First run - Clean and run all the scripts
On setting up the server, run first.py. It does the following
a. Cleans up gen folder if present
b. Logins to TPO portal and saves the html file in gen
c. Extracts all the notices from this html and saves to gen/json folder
d. Update all the json files in gen/json folder to include details and attachments
e. Change the 'sent' attribute of these json files to False - So that old
   notifications are not sent to the channel.
f. Set up cron jobs to automate the process.

Signed-off-by: Ravi Ranjan <ravirnjn88@gmail.com>