Pythonic way of getters and setter using poperty
Python is not Java (http://dirtsimple.org/2004/12/python-is-not-java.html)
So, java way of getters and setters are not applicable here.
If one really want to have getters and setters in python, property decorator
should be used as is done in this commit.

Further, these property decorators are not valid for old class python classes
but only new ones. Hence the class TpoSession should inherit from 'object'
class.

Signed-off-by: Ravi Ranjan <ravirnjn88@gmail.com>