Beginning logging of application #7
As my experience with various bugs suggest, it is extremely important to have
logging in your application. For this, I have started integrating python logging
module with this application.

This commit represents the initial stab at logging integration.
Logs are sent to /logs/date.log file and are of DEBUG level giving us a huge
amount of logs to study.

Signed-off-by: Ravi Ranjan <ravirnjn88@gmail.com>