Use postgresql database instead of sqlite
sqlite saves the database in local storage as .db files. This system presents
the same difficulties on heroku where the local storage is ephemeral.
A better option is postgresql which is natively supported by heroku.
To use notice_db, one has to describe an environmental variable DATABASE_URL
pointing to the postgres link of current user.
Do this by adding the following in the .bashrc

>  export DATABASE_URL=postgres:///$(whoami)

Also, psycopg2 is required to interact with postgres in python

Signed-off-by: Ravi Ranjan <ravirnjn88@gmail.com>