Update notice in db to include text Attachments database made
Using unimplemented methods in the NoticeWrapper class, perform updation of
notices in the database.

Attachments needs a database of itself. But since a notice can contain many
attachments, it is made in form of a linked list using self referential foreign
key.

Notice adds a foreign key to point to first attachment,

Signed-off-by: Ravi Ranjan <ravirnjn88@gmail.com>