Send unsent notices from the database
Not, send unsent notices from the database to the pushbullet channel.
For this, common parts, like sending payload to pushbullet has been separated
into different function which is used both by database and json functions to do
the sending. Added the function to NoticeWrapper as well. But it remains
unimplemented.

Signed-off-by: Ravi Ranjan <ravirnjn88@gmail.com>