Implement insert_dict static method of NoticeWrapper
NoticeWrapper provides the interface to interact with Notice database.
Give a dictionary with appropriate parameters to insert_dict method and
NoticeWrapper adds it to the database.

But it is better to use insert_dict_safe method which first looks to see if
Notice with the same title is already present. If yes, then return immediatedly.
If not, insert a new notice to database.

For the implementation of insert_dict method, classmethods from_dict and
from_list have been added to Notice and Attachment.

Insert script has been modified to use insert_dict_safe. Also, if argument db is
provided, insert notices to db not saving locally.

Signed-off-by: Ravi Ranjan <ravirnjn88@gmail.com>